AB#2977 -- Refactoring to add and use custom useHCaptcha hook and route helper
