+++ frontend/app/utils/hcaptcha-utils.ts
@@ -1,0 +1,25 @@
+import { useEffect, useRef } from 'react';
+
+import HCaptcha from '@hcaptcha/react-hcaptcha';
+
+export function useHCaptcha() {
+  const captchaRef = useRef<HCaptcha>(null);
+
+  useEffect(() => {
+    let timeoutId: ReturnType<typeof setTimeout>;
+
+    if (captchaRef.current?.isReady()) {
+      captchaRef.current.execute();
+    } else {
+      timeoutId = setTimeout(() => {
+        captchaRef.current?.execute();
+      }, 500);
+    }
+
+    return () => {
+      clearTimeout(timeoutId);
+    };
+  }, []);
+
+  return { captchaRef };
+}
